U2
Weapon system added, Fleet system starting

+++ Damage.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 public class Damage
 	{
 		public float damage;
-		public int penetration;
+		public float penetration;
 		public Vector3 point;
 
-		public Damage (float dmg, int pnt, Vector3 place)
+		public Damage (float dmg, float pnt, Vector3 place)
 		{
 			if (dmg < 0) damage= 0; else damage = dmg;
 			if (pnt < 0) penetration = 0; else penetration = pnt;
